+++ wreport/buffers/bufr.cc
@@ -5,6 +5,7 @@
 
 // #define TRACE_INTERPRETER
 
+/*
 #ifdef TRACE_INTERPRETER
 #define TRACE(...) fprintf(stderr, __VA_ARGS__)
 #define IFTRACE if (1)
@@ -12,6 +13,7 @@
 #define TRACE(...) do { } while (0)
 #define IFTRACE if (0)
 #endif
+*/
 
 using namespace std;