+++ wreport/bulletin/dds-validator.cc
@@ -1,25 +1,3 @@
-/*
- * bulletin/dds-validator - Validate variables of a bulletin against its data
- *                          descriptor section
- *
- * Copyright (C) 2011--2015  ARPA-SIM <urpsim@smr.arpa.emr.it>
- *
- * This program is free software; you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation; either version 2 of the License.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program; if not, write to the Free Software
- * Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA  02110-1301 USA
- *
- * Author: Enrico Zini <enrico@enricozini.com>
- */
-
 #include "wreport/bulletin/dds-validator.h"
 #include "wreport/vartable.h"
 
@@ -60,7 +38,7 @@ void DDSValidator::check_fits(Varinfo info, const Var& var)
             if (!is_crex)
             {
                 if (encoded >= (1u<<info->bit_len))
-                    error_consistency::throwf("value %f (%u) does not fit in %d bits", var.enqd(), encoded, info->bit_len);
+                    error_consistency::throwf("value %f (%u) does not fit in %u bits", var.enqd(), encoded, info->bit_len);
             }
         }
         break;