+++ NEWS.md
@@ -1,3 +1,17 @@
+# New in version 3.38
+
+* Updated code to use C++17 features
+* Building against wreport requires C++17 from now on
+* API and ABI are kept consistent, except for `wreport/utils/tests.h`
+
+# New in version 3.37
+
+* Reinclude table v. 33 which was excluded with meson
+
+# New in version 3.36
+
+* Include cstdint in wreport/utils/string.cc (#55)
+
 # New in version 3.35
 
 * Fixed a corner case in decoding of associated fields (#52)